requirements.txt
markitdown[all]
python-dotenv
requests
groq
openai
fitz
deepl
pymupdf
streamlit
PyQt6